refactor: rename Pagination component to Paginator component
re #292
