+++ pages/join/public/script.js
@@ -1,10 +1,10 @@
 import Button from '~/components/button/index.vue'
-import Pagination from '~/components/pagination/index.vue'
+import Paginator from '~/components/paginator/index.vue'
 
 export default {
   components: {
     Button,
-    Pagination,
+    Paginator,
   },
   layout: 'after-login/index',
   data() {
@@ -12,7 +12,6 @@ export default {
       page: 1,
       limit: 10,
       total: 0,
-      pages: [],
       worlds: [],
       error: false,
       errorMessage: 'さんかにしっぱいしました',
@@ -31,28 +30,6 @@ export default {
         this.page = res.page
         this.total = res.total
         this.worlds = res.list
-        this.pagination()
-      }
-    },
-    pagination() {
-      this.pages = []
-      const pages = Math.ceil(this.total / this.limit)
-      let min = 0
-      let max = 0
-      if (this.page <= 3) {
-        min = 1
-        max = pages <= 5 ? pages : 5
-      } else if (pages < this.page + 2) {
-        if (this.page <= 5) min = 1
-        else min = pages === this.page ? this.page - 4 : this.page - 3
-        max = pages
-      } else {
-        min = this.page - 2
-        max = this.page + 2
-      }
-      for (let i = 0; i < 5; i++) {
-        const num = min + i
-        if (num <= max) this.pages.push(num)
       }
     },
     async joinWorld(worldId) {
