+++ pages/join/public/template.html
@@ -8,7 +8,7 @@
       </tr>
     </thead>
     <tbody class="table-body">
-      <tr v-for="world in worlds" v-bind:key="world.worldId" class="table-body__row">
+      <tr v-for="world in worlds" :key="world.worldId" class="table-body__row">
         <td class="table-body__col">{{ world.worldId }}</td>
         <td class="table-body__col">{{ world.role === 1 ? 'うさぎさん' : 'ばいきんくん' }}</td>
         <td class="table-body__col">
@@ -21,12 +21,12 @@
       </tr>
     </tbody> 
   </table>
-  <Pagination
+  <p v-if="!total" class="not-world">さんかできるワールドがありません</p>
+  <Paginator
     class="pagination"
     :page="page"
     :limit="limit"
     :total="total"
-    :pages="pages"
     @search="searchWorld"
   />
   <p v-if="error" class="error-message">{{ errorMessage }}</p>
